support redeem energy point

+++ oas-server/oas/src/main/java/com/cascv/oas/server/blockchain/wrapper/EnergyPointRedeem.java
@@ -1,6 +1,7 @@
 package com.cascv.oas.server.blockchain.wrapper;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -10,7 +11,7 @@ public class EnergyPointRedeem implements Serializable {
   private static final long serialVersionUID = 1L;
   @Setter @Getter private String date;
   @Setter @Getter private Integer type;
-  @Setter @Getter private Double rate;
-  @Setter @Getter private Integer amount;
+  @Setter @Getter private BigDecimal rate;
+  @Setter @Getter private BigDecimal amount;
     
 }
